+++ NEWS.md
@@ -1,6 +1,18 @@
 
 # pkgcache, development version
 
+## BREAKING CHANGE
+
+* Starting from version 2.0.0 pkgcache returns data frames instead
+  of tibbles. While data frames and tibbles are very similar, they are
+  not completely compatible. To convert the output of pkgcache to tibbles
+  call the `as_tibble()` function on them.
+
+  pkgcache loads the pillar package at startup, if available, and uses it
+  to improve the printing of pkgcache data frames.
+
+## OTHER CHANGES
+
 * pkgcache now uses HTTP 1.1 on Linux as well, because of crashes with
   HTTP/2.